requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
pandas==2.2.3
Werkzeug==3.0.4
PyJWT==2.8.0
gunicorn==23.0.0